rename: 컴포넌트 이름 kebab case로 변경
