+++ app/(routes)/isa/page.tsx
@@ -1,4 +1,4 @@
-import ISAPageContainer from './_components/ISAPageContainer';
+import ISAPageContainer from './_components/isa-page-container';
 
 const ISAPage = () => {
   return <ISAPageContainer />;
